Abstract local-storage functionality
